feature/52130 = refactor code 2
