+++ angular/src/bsb-theme/css/style.css
@@ -10579,26 +10579,11 @@ thead{
   background: #fff;
 }
 
-.preReview::-webkit-scrollbar
-{
-	width: 7px;
+.custom-scrollbar::-webkit-scrollbar {
+  width: 7px;
 	background-color: #F5F5F5;
 }
-
-.preReview::-webkit-scrollbar-thumb
-{
-	background-color: #cecece;
-	border: 2px solid #cecece;
-}
-
-.review::-webkit-scrollbar
-{
-	width: 7px;
-	background-color: #F5F5F5;
-}
-
-.review::-webkit-scrollbar-thumb
-{
-	background-color: #cecece;
+.custom-scrollbar::-webkit-scrollbar-thumb {
+  background-color: #cecece;
 	border: 2px solid #cecece;
 }
